Extract providerLabel to individual component
