+++ assets/js/components/ProviderLabel/index.js
@@ -1,0 +1,3 @@
+import ProviderLabel from './ProviderLabel';
+
+export default ProviderLabel;
